chore: tidy of things that were where they shouldnt, precise purpose for each file
