+++ test/e2e/page-objects/connect-wallet.ts
@@ -1,5 +1,5 @@
 import { By, WebDriver } from 'selenium-webdriver'
-import { clickElement, getByDataTestID, isElementDisplayed, waitForElementToDisappear } from '../selenium-util'
+import { clickElement, getByDataTestID, isElementDisplayed, waitForElementToDisappear } from '../helpers/selenium-util'
 import locators from '../../../frontend/components/locators'
 
 export class ConnectWallet {
