+++ test/e2e/settings.spec.ts
@@ -1,9 +1,9 @@
 import { WebDriver } from 'selenium-webdriver'
-import { captureScreenshot } from './driver'
-import { navigateToExtensionLandingPage, createWalletAndDriver } from './wallet-helpers/common'
+import { captureScreenshot } from './helpers/driver'
+import { navigateToExtensionLandingPage, createWalletAndDriver } from './helpers/wallet/common-wallet-values'
 import { NavPanel } from './page-objects/navpanel'
-import { switchWindowHandles } from './selenium-util'
-import { VegaAPI } from './wallet-helpers/vega-api'
+import { switchWindowHandles } from './helpers/selenium-util'
+import { VegaAPI } from './helpers/wallet/vega-api'
 import { Transaction } from './page-objects/transaction'
 import { ConnectWallet } from './page-objects/connect-wallet'
 import { Settings } from './page-objects/settings'
